+++ src/vite-env.d.ts
@@ -4,3 +4,7 @@ declare module '*.vue' {
     const component: DefineComponent<object, object, any>;
     export default component;
 }
+
+interface Window {
+    __STK_TB_ID_COUNT__: number;
+}